Adjust to use microseconds write on servo

+++ firmware/libraries/ros_lib/kobuki_msgs/DigitalInputEvent.h
@@ -22,7 +22,7 @@ namespace kobuki_msgs
     virtual int serialize(unsigned char *outbuffer) const
     {
       int offset = 0;
-      for( uint8_t i = 0; i < 4; i++){
+      for( uint32_t i = 0; i < 4; i++){
       union {
         bool real;
         uint8_t base;
@@ -37,7 +37,7 @@ namespace kobuki_msgs
     virtual int deserialize(unsigned char *inbuffer)
     {
       int offset = 0;
-      for( uint8_t i = 0; i < 4; i++){
+      for( uint32_t i = 0; i < 4; i++){
       union {
         bool real;
         uint8_t base;
